Retrieve strategy from data analysis
