+++ cg/services/delivery_message/messages/fastq_message.py
@@ -1,10 +1,10 @@
 from cg.services.delivery_message.messages import DeliveryMessage
-from cg.services.delivery_message.utils import get_fastq_delivery_path
+from cg.services.delivery_message.messages.utils import get_fastq_delivery_path
 from cg.store.models import Case
 
 
 class FastqMessage(DeliveryMessage):
-    def generate_message(self, case: Case):
+    def create_message(self, case: Case):
         delivery_path: str = get_fastq_delivery_path(case)
         return (
             f"Hello,\n\n "
